guardado de detalles de manera asincronica (EventDetail); refactorizacion de codigo
