+++ src/main/java/ar/edu/iw3/integration/cli2/controllers/OrderCli2RestController.java
@@ -1,7 +1,7 @@
 package ar.edu.iw3.integration.cli2.controllers;
 
 import ar.edu.iw3.controllers.Constants;
-import ar.edu.iw3.integration.cli2.model.business.OrderCli2Business;
+import ar.edu.iw3.integration.cli2.model.business.implementaions.OrderCli2Business;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
@@ -32,7 +32,7 @@ public class OrderCli2RestController {
             @RequestHeader("orderNumber") String orderNumber,
             @RequestHeader("finalWeight") float initialWeight) {
         HttpHeaders headers = new HttpHeaders();
-        headers.set(HttpHeaders.CONTENT_TYPE, "application/pdf");
+        headers.set(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_PDF_VALUE);
         headers.set(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"report.pdf\"");
         byte[] pdfContent = orderCli2Business.registerFinalWeighing(orderNumber, initialWeight);
         return new ResponseEntity<>(pdfContent, headers, HttpStatus.OK);
