+++ src/main/java/ar/edu/iw3/integration/cli3/controllers/OrderCli3RestController.java
@@ -1,19 +1,12 @@
 package ar.edu.iw3.integration.cli3.controllers;
 
 import ar.edu.iw3.controllers.Constants;
-
-import ar.edu.iw3.integration.PdfService;
 import ar.edu.iw3.integration.cli3.OrderCli3SlimV1JsonSerializer;
-
-import ar.edu.iw3.integration.cli3.model.business.IDetailCli3Business;
-import ar.edu.iw3.integration.cli3.model.business.IOrderCli3Business;
+import ar.edu.iw3.integration.cli3.model.business.interfaces.IOrderCli3Business;
 import ar.edu.iw3.model.Detail;
 import ar.edu.iw3.model.Order;
-
 import ar.edu.iw3.util.JsonUtils;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import com.itextpdf.text.DocumentException;
-
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -28,6 +21,10 @@ public class OrderCli3RestController {
     @Autowired
     private IOrderCli3Business orderCli3Business;
 
+    @Autowired
+    private IOrderCli3Business orderBusiness;
+    ;
+
     // Endpoint para validar password y obtener id de la orden y preset de carga
     // todo pasar password por header?
     @SneakyThrows
@@ -36,8 +33,8 @@ public class OrderCli3RestController {
         Order order = orderCli3Business.validatePassword(Integer.parseInt(password));
 
         // Serializacion del resultado
-        StdSerializer<Order> ser = new OrderCli3SlimV1JsonSerializer(Order.class,false);
-        String result = JsonUtils.getObjectMapper(Order.class,ser, null).writeValueAsString(order);
+        StdSerializer<Order> ser = new OrderCli3SlimV1JsonSerializer(Order.class, false);
+        String result = JsonUtils.getObjectMapper(Order.class, ser, null).writeValueAsString(order);
 
         return new ResponseEntity<>(result, HttpStatus.OK);
     }
@@ -51,16 +48,12 @@ public class OrderCli3RestController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @Autowired
-    private IDetailCli3Business detailBusiness;
-
     // Endpoint para recibir datos de carga
     @SneakyThrows
     @PostMapping("/detail")
     public ResponseEntity<?> receiveLoadData(@RequestBody Detail detail) {
-        detailBusiness.receiveDetails(detail);
+        orderBusiness.receiveDetails(detail);
         return new ResponseEntity<>(HttpStatus.OK);
     }
-
 
 }
